% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_matrix.R
\name{apply_matrix}
\alias{apply_matrix}
\title{apply any matrix transformation to a set of points}
\usage{
apply_matrix(pointset, transformation_matrix, center_of_action = NULL)
}
\arguments{
\item{pointset}{a n-by-2 matrix of n 2D points}

\item{transformation_matrix}{a 2-by-2 transformation matrix}

\item{center_of_action}{an optional origin point (2D position vector)
from which to apply the transformation}
}
\value{
a transformed pointset
}
\description{
optional: given center of action (default: centroid)
}
