% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pca.R
\name{apply_pca}
\alias{apply_pca}
\title{Transform points into Principal Component space}
\usage{
apply_pca(pointset, pca)
}
\arguments{
\item{pointset}{a n-by-2 matrix of n 2D points}

\item{pca}{the result of PCA computation (list)}
}
\value{
a transformed pointset
}
\description{
Transform points into Principal Component space
}
