% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rotation.R
\name{apply_rotation}
\alias{apply_rotation}
\title{apply a rotation matrix to a set of points}
\usage{
apply_rotation(pointset, rotation_matrix, center_of_action = NULL)
}
\description{
around a given center of rotation (default: centroid)
}
