% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_scaling.R
\name{apply_scaling}
\alias{apply_scaling}
\title{scale a set of point by a factor (without moving)}
\usage{
apply_scaling(pointset, scaling, axis = NULL, fixpoint = NULL)
}
\arguments{
\item{pointset}{a n-by-2 matrix of n 2D points}

\item{scaling}{a scalar by which to scale the points}

\item{axis}{the coordinate axis along which to scale (anisotropic scaling)}

\item{fixpoint}{an optional origin point (2D position vector)
from which to apply scaling}
}
\value{
a transformed pointset
}
\description{
scaling has to happen around the points' centroid
optional anisotropic scaling: choose one axis
optionally set a fixpoint (for scaling off-centroid)
}
