% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_shear.R
\name{apply_shear}
\alias{apply_shear}
\title{apply a shear matrix to a set of points}
\usage{
apply_shear(pointset, shear_matrix, center_of_action = NULL)
}
\description{
around a given center of shearing (default: centroid)
}
