% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_centroid.R
\name{compute_centroid}
\alias{compute_centroid}
\title{the centroid of a set of points}
\usage{
compute_centroid(pointset)
}
\arguments{
\item{pointset}{a n-by-2 matrix of n 2D points}
}
\value{
the centroid as 2D position vector
}
\description{
the centroid of a set of points
}
