% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pca.R
\name{compute_pca}
\alias{compute_pca}
\title{Compute Principal Components (via Eigenanalysis) of a point matrix}
\usage{
compute_pca(pointset)
}
\arguments{
\item{pointset}{a n-by-2 matrix of n 2D points}
}
\value{
list with PCA results
}
\description{
Compute Principal Components (via Eigenanalysis) of a point matrix
}
