% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rcs.R
\name{compute_rcs}
\alias{compute_rcs}
\title{compute root centroid size (RCS), i.e. the scale of the data}
\usage{
compute_rcs(pointset, axis = NULL)
}
\description{
RCS is a scalar, defined as the root sum of
euclidean distances of all points from their centroid.
Optionally calculate size along only single axis.
}
