% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shear_cartesian.R
\name{compute_shear_cartesian}
\alias{compute_shear_cartesian}
\title{calculate the actual shear of a pointset relative to reference}
\usage{
compute_shear_cartesian(
  pointset,
  reference_pointset,
  fix_x = FALSE,
  fix_y = FALSE
)
}
\arguments{
\item{pointset}{a n-by-2 matrix of n 2D points}

\item{reference_pointset}{another 2D pontset (same n-by-2 dimensions)
which serves as reference}

\item{fix_x}{do not shear along the first dimension}

\item{fix_y}{do not shear along the second dimension}
}
\value{
relative shear matrix of the pointset to the reference
}
\description{
optionally fixing either direction
}
