% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{deg2rad}
\alias{deg2rad}
\title{angles: degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{angle in degrees \verb{[0, 360]}}
}
\value{
angle in radians \verb{[0, 2π]}
}
\description{
angles: degrees to radians
}
