% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pca.R
\name{inverse_transform_pca}
\alias{inverse_transform_pca}
\title{Transform points back from Principal Component space}
\usage{
inverse_transform_pca(pointset, pca)
}
\arguments{
\item{pointset}{a n-by-m matrix of principal components}

\item{pca}{the result of PCA computation (list)}
}
\value{
a pointset transformed back to the original space
}
\description{
Transform points back from Principal Component space
}
