% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{invert_rotation_matrix}
\alias{invert_rotation_matrix}
\title{invert a rotation matrix}
\usage{
invert_rotation_matrix(M)
}
\arguments{
\item{M}{rotation matrix}
}
\value{
inverse rotation matrix
}
\description{
invert a rotation matrix
}
