% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{invert_scaling}
\alias{invert_scaling}
\title{invert a scaling factor}
\usage{
invert_scaling(s)
}
\arguments{
\item{s}{scaling factor}
}
\value{
inverse scaling factor
}
\description{
invert a scaling factor
}
