% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{invert_shear_matrix}
\alias{invert_shear_matrix}
\title{invert a shear operation}
\usage{
invert_shear_matrix(M)
}
\arguments{
\item{M}{shear matrix}
}
\value{
inverse shear matrix
}
\description{
invert a shear operation
}
