% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{invert_translation}
\alias{invert_translation}
\title{invert a translation}
\usage{
invert_translation(v)
}
\arguments{
\item{v}{translation vector}
}
\value{
inverse translation vector
}
\description{
invert a translation
}
