% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_square_cartesian.R
\name{make_square_cartesian}
\alias{make_square_cartesian}
\title{apply scaling for anisotropic correction (x-RCS to be the same as y-RCS)}
\usage{
make_square_cartesian(pointset, aspect_ratio = NULL, axis = 1)
}
\arguments{
\item{pointset}{a n-by-2 matrix of n 2D points}

\item{aspect_ratio}{the aspect ratio by which the
existing aspect is corrected (divided)}

\item{axis}{the coordinate axis along which to scale (anisotropic scaling)}
}
\value{
a transformed pointset
}
\description{
this does not literally create a square, but corrects the point sets aspect.
}
