% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_translation.R
\name{move}
\alias{move}
\title{alias for \code{apply_translation}}
\usage{
move(pointset, shift)
}
\arguments{
\item{pointset}{a n-by-2 matrix of n 2D points}

\item{shift}{a 2D vector (directed translation offset)
along which points are moved}
}
\description{
alias for \code{apply_translation}
}
