% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{plot_points}
\alias{plot_points}
\title{quickly get an overview of the points of interest}
\usage{
plot_points(points, ref_points = NULL, ...)
}
\description{
quickly get an overview of the points of interest
}
