% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes_superimposition.R
\name{procrustes_superimposition}
\alias{procrustes_superimposition}
\title{calculate Procrustes transformation parameters}
\usage{
procrustes_superimposition(
  points_interest,
  points_ref,
  disable_mirror = TRUE,
  mirror_x = FALSE
)
}
\arguments{
\item{points_interest}{a n-by-2 matrix of n 2D points which is
the shape to be transformed}

\item{points_ref}{a n-by-2 matrix of n 2D points which serves
as the reference shape}

\item{disable_mirror}{avoid mirroring during rotation}

\item{mirror_x}{enforce prior mirroring of x-values on the y-axis}
}
\value{
a list containing the calculated transformation
operations/parameters and Procrustes Distance
}
\description{
calculate Procrustes transformation parameters
}
