% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{rad2deg}
\alias{rad2deg}
\title{angles: radians to degrees}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{angle in radians \verb{[0, 2π]}}
}
\value{
angle in degrees \verb{[0, 360]}
}
\description{
angles: radians to degrees
}
