% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{rotation_matrix_from_angle_deg}
\alias{rotation_matrix_from_angle_deg}
\title{compose a rotation matrix from an angle (degrees!)}
\usage{
rotation_matrix_from_angle_deg(deg)
}
\arguments{
\item{deg}{angle in degrees \verb{[0, 360]}}
}
\value{
2-by-2 rotation matrix
}
\description{
compose a rotation matrix from an angle (degrees!)
}
