% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{rotation_matrix_from_angle_rad}
\alias{rotation_matrix_from_angle_rad}
\title{compose a rotation matrix from an angle (radians!)}
\usage{
rotation_matrix_from_angle_rad(rad)
}
\arguments{
\item{rad}{angle in radians \verb{[0, 2π]}}
}
\value{
2-by-2 rotation matrix
}
\description{
compose a rotation matrix from an angle (radians!)
}
