% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_shear.R
\name{shear}
\alias{shear}
\title{alias for \code{apply_shear}}
\usage{
shear(pointset, shear_matrix, center_of_action = NULL)
}
\arguments{
\item{pointset}{a n-by-2 matrix of n 2D points}

\item{shear_matrix}{a 2-by-2 shear matrix matrix}

\item{center_of_action}{an optional origin point (2D position vector)
from which to apply the shear transformation}
}
\description{
alias for \code{apply_shear}
}
