% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{standardize a pointset (options: translation, scaling, rotation)}
\usage{
standardize(
  pointset,
  fix_position = FALSE,
  fix_size = FALSE,
  fix_rotation = TRUE
)
}
\arguments{
\item{pointset}{a n-by-2 matrix of n 2D points}

\item{fix_position}{whether or not to apply translation}

\item{fix_size}{whether or not to apply scaling}

\item{fix_rotation}{whether or not to apply rotation}
}
\value{
a transformed pointset
}
\description{
affine transformations to default values
}
