% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_matrix.R
\name{apply_matrix}
\alias{apply_matrix}
\title{apply any matrix transformation to a set of points}
\usage{
apply_matrix(pointset, transformation_matrix, center_of_action = NULL)
}
\description{
optional: given center of action (default: centroid)
}
