% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rotation.R
\name{apply_rotation}
\alias{apply_rotation}
\title{apply a rotation matrix to a set of points}
\usage{
apply_rotation(pointset, rotation_matrix, center_of_action = NULL)
}
\arguments{
\item{pointset}{a n-by-2 matrix of n 2D points}

\item{rotation_matrix}{a 2-by-2 rotation matrix}

\item{center_of_action}{an optional origin point (2D position vector)
from which to apply rotation}
}
\value{
a transformed pointset
}
\description{
around a given center of rotation (default: centroid)
}
