% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_scaling.R
\name{apply_scaling}
\alias{apply_scaling}
\title{scale a set of point by a factor (without moving)}
\usage{
apply_scaling(pointset, scaling, axis = NULL, fixpoint = NULL)
}
\description{
scaling has to happen around the points' centroid
optional anisotropic scaling: choose one axis
optionally set a fixpoint (for scaling off-centroid)
}
