% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes_superimposition.R
\name{apply_transformations}
\alias{apply_transformations}
\title{This helper function will apply a previously calculated series of transformations to data.}
\usage{
apply_transformations(pointset, params)
}
\description{
Actions include, in order of appearance:
\itemize{
\item forced mirror x values on the y-axis
\item centering to the centroid
\item rotation by the rotation matrix
\item scaling with root centroid size
\item translation by a shift vector
}
}
