% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aspects_cartesian.R
\name{compute_aspects_cartesian}
\alias{compute_aspects_cartesian}
\title{compute the aspect ratios of a point set, relative to one axis}
\usage{
compute_aspects_cartesian(pointset, axis = 1)
}
\description{
compute ratios of the centroid size axes relative to one focal axis
}
