% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rcs.R
\name{compute_rcs}
\alias{compute_rcs}
\title{compute root centroid size (RCS), i.e. the scale of the data}
\usage{
compute_rcs(pointset, axis = NULL)
}
\arguments{
\item{pointset}{a n-by-2 matrix of n 2D points}

\item{axis}{the coordinate axis along which to calculate}
}
\value{
RCS = Root Centroid Size of the pointset
}
\description{
RCS is a scalar, defined as the root sum of
euclidean distances of all points from their centroid.
Optionally calculate size along only single axis.
}
