% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rotation.R
\name{compute_rotation}
\alias{compute_rotation}
\title{Apply Kabsch algorithm to find rotation to reference}
\usage{
compute_rotation(pointset, reference_pointset, prevent_mirroring = TRUE)
}
\description{
Apply Kabsch algorithm to find rotation to reference
}
