% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shear_cartesian.R
\name{compute_shear_cartesian}
\alias{compute_shear_cartesian}
\title{calculate the actual shear of a pointset relative to reference}
\usage{
compute_shear_cartesian(
  pointset,
  reference_pointset,
  fix_x = FALSE,
  fix_y = FALSE
)
}
\description{
optionally fixing either direction
}
