% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aspects_cartesian.R
\name{compute_xy_aspect_cartesian}
\alias{compute_xy_aspect_cartesian}
\title{specifically compute the x/y aspect along cartesian axes}
\usage{
compute_xy_aspect_cartesian(ps)
}
\description{
specifically compute the x/y aspect along cartesian axes
}
