% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_square_cartesian.R
\name{make_square_cartesian}
\alias{make_square_cartesian}
\title{apply scaling for anisotropic correction (x-RCS to be the same as y-RCS)}
\usage{
make_square_cartesian(pointset, aspect_ratio = NULL, axis = 1)
}
\description{
this does not literally create a square, but corrects the point sets aspect.
}
