% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes_superimposition.R
\name{procrustes_superimposition}
\alias{procrustes_superimposition}
\title{calculate Procrustes transformation parameters}
\usage{
procrustes_superimposition(
  points_interest,
  points_ref,
  disable_mirror = TRUE,
  mirror_x = FALSE
)
}
\description{
calculate Procrustes transformation parameters
}
