% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{rotation_matrix_from_angle_deg}
\alias{rotation_matrix_from_angle_deg}
\title{compose a rotation matrix from an angle (degrees!)}
\usage{
rotation_matrix_from_angle_deg(deg)
}
\description{
compose a rotation matrix from an angle (degrees!)
}
