% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_to_size.R
\name{scale_to_size}
\alias{scale_to_size}
\title{Scale the points so that their RCS matches a target size}
\usage{
scale_to_size(pointset, target = NULL, axis = NULL, fixpoint = NULL)
}
\description{
By default, pointset is scaled to unit RCS.
Optionally scale only a single axis,
or relative to a fixpoint (which will change centroid position)
}
