% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_to_point.R
\name{shift_to_point}
\alias{shift_to_point}
\title{move pointset so that its centroid matches a target point}
\usage{
shift_to_point(pointset, target = NULL)
}
\arguments{
\item{pointset}{a n-by-2 matrix of n 2D points}

\item{target}{a 2D position vector to become the new centroid}
}
\value{
a transformed pointset
}
\description{
Default target point is the coordinate origin
}
