% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{standardize a pointset (options: translation, scaling, rotation)}
\usage{
standardize(
  pointset,
  fix_position = FALSE,
  fix_size = FALSE,
  fix_rotation = TRUE
)
}
\description{
affine transformations to default values
}
